Remove files, rename all the linters
