--- test/linters/projects/php-codesniffer/file1.php
@@ -1,8 +1,0 @@
-<?php
-
-function helloWorld()
-{
-    echo 'Hello World';
-}
-
-helloWorld(); // "PSR1.Files.SideEffects.FoundWithSymbol" warning
