--- test/linters/projects/php-codesniffer/file2.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Vendor\Package;
-
-class ClassName { // "Squiz.Functions.MultiLineFunctionDeclaration.ContentAfterBrace" error
-    private $foo = null;
-}
-
-// "PSR2.Files.ClosingTag.NotAllowed" error
-?>
