--- test/linters/projects/php-codesniffer/phpcs.xml.dist
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<ruleset name="lint-action-test">
-	<file>.</file>
-	<arg name="basepath" value="." />
-	<rule ref="PSR2"/>
-</ruleset>
