--- test/linters/projects/prettier/.prettierrc.json
@@ -1,5 +1,0 @@
-{
-	"semi": true,
-	"singleQuote": true,
-	"useTabs": true
-}
