--- test/linters/projects/prettier/file1.js
@@ -1,5 +1,0 @@
-function main() {
-	console.log("Hello world"); // "singleQuote" error
-}
-
-main() // "semi" error
