--- test/linters/projects/prettier/file2.css
@@ -1,1 +1,0 @@
-body {color: red;} /* Line break errors */
